Add support for unit tests
Add support for gradle to run unit test
