+++ build.gradle
@@ -45,4 +45,5 @@ checkstyle {
 
 run{
     standardInput = System.in
+    enableAssertions = true
 }
